requirements.txt
Django==4.2
django-cron
django-timezone-field
djangorestframework
Pillow
django-cors-headers
leaflet
drf_spectacular
django-leaflet
gunicorn
python-dotenv